scing-voice-interface/requirements.txt
openai-whisper
sounddevice
numpy
pyttsx3
openai
python-dotenv
